requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.8.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
